+++ native/iosTest/Pods/Target Support Files/Pods-WatermelonTesterTests/Pods-WatermelonTesterTests-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Pods_WatermelonTesterTests : NSObject
+@end
+@implementation PodsDummy_Pods_WatermelonTesterTests
+@end