+++ native/iosTest/WatermelonTester/WatermelonTester-Bridging-Header.h
@@ -1,0 +1,4 @@
+#import <React/RCTBundleURLProvider.h>
+#import <React/RCTRootView.h>
+#import <React/RCTViewManager.h>
+#import <React/RCTBridgeModule.h>