+++ src/observation/encodeMatcher/test.js
@@ -1,9 +1,8 @@
 import Query from 'Query'
 import * as Q from 'QueryDescription'
 import Model from 'Model'
-import { matchTests } from '__tests__/databaseTests'
-
 import encodeMatcher from 'observation/encodeMatcher'
+import { matchTests } from '../../__tests__/databaseTests'
 
 const mockModelClass = { table: 'tasks' }
 const mockCollection = { modelClass: mockModelClass }